requirements.txt
tensorflow>=2.8.0
flask>=2.0.1
pillow>=8.3.1
numpy>=1.19.5
werkzeug>=2.0.1
scikit-learn>=1.0.0
matplotlib>=3.4.3
pandas>=1.3.3